+++ src/main/java/de/johni0702/minecraft/bobby/mixin/WorldRendererMixin.java
@@ -1,0 +1,14 @@
+package de.johni0702.minecraft.bobby.mixin;
+
+import net.minecraft.client.render.WorldRenderer;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.ModifyArg;
+
+@Mixin(WorldRenderer.class)
+public abstract class WorldRendererMixin {
+    @ModifyArg(method = "method_37365", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/render/BackgroundRenderer;applyFog(Lnet/minecraft/client/render/Camera;Lnet/minecraft/client/render/BackgroundRenderer$FogType;FZF)V"), index = 2)
+    private static float clampMaxValue(float viewDistance) {
+        return Math.min(viewDistance, 32 * 16);
+    }
+}